feat: add RadioCore e-paper support

+++ src/Displays/BaseDisplay/graphical.cpp
@@ -51,7 +51,7 @@ void BaseDisplay::drawPixel(int16_t x, int16_t y, uint16_t color) {
     // Check if pixel falls in our page
     if((uint16_t) x >= winrot_left && (uint16_t) y >= page_top && (uint16_t) y <= page_bottom && (uint16_t) x <= winrot_right) {
 
-        uint16_t byte_offset;
+        uint32_t byte_offset;
         uint8_t bit_offset;
         calculatePixelPageOffset(x, y, byte_offset, bit_offset);    // Position of pixel within the page files. Overriden if no "partial window" support
 
@@ -69,9 +69,9 @@ void BaseDisplay::drawPixel(int16_t x, int16_t y, uint16_t color) {
 }
 
 // Where should the pixel be placed in the pagefile - overriden by derived display classes which do not support "partial window"
-void BaseDisplay::calculatePixelPageOffset(uint16_t x, uint16_t y, uint16_t &byte_offset, uint8_t &bit_offset) {
+void BaseDisplay::calculatePixelPageOffset(uint16_t x, uint16_t y, uint32_t &byte_offset, uint8_t &bit_offset) {
     // Calculate a memory location (byte) for our pixel
-    byte_offset = (y - page_top) * ((winrot_right - winrot_left + 1) / 8);
+    byte_offset = (uint32_t) (y - page_top) * ((winrot_right - winrot_left + 1) / 8);
     byte_offset += ((x - winrot_left) / 8);     
     bit_offset = x % 8;   // Find the location of the bit in which the value will be stored
     bit_offset = (7 - bit_offset);  // For some reason, the screen wants the bit order flipped. MSB vs LSB?
@@ -105,7 +105,7 @@ bool BaseDisplay::supportsColor(Color c) {
     // Pixels are modified the in MCU's memory. Requires update() to draw the changes to display.
     void BaseDisplay::invert() {
         // Invert all pixels of the existing image
-        for(uint16_t i = 0; i < page_bytecount; i++)
+        for(uint32_t i = 0; i < page_bytecount; i++)
         page_black[i] = page_black[i] ^ 0xFF;
     }
 
@@ -128,7 +128,7 @@ bool BaseDisplay::supportsColor(Color c) {
         page_black = original_image;
 
         // Invert any pixels of the original image which line up with a 1 in the mask
-        for(uint16_t i = 0; i < page_bytecount; i++)
+        for(uint32_t i = 0; i < page_bytecount; i++)
             page_black[i] = page_black[i] ^ mask[i];
 
         // Free the memory used by the mask
